backend/requirements.txt
asgiref==3.7.2
Django==3.2.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
djoser==2.1.0
django-filter==23.2
djangorestframework==3.12.4
Pillow==10.0.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
gunicorn==20.1.0
